Made a number of modifications:
Added comments to all functions
Deleted unecessary comments
Updated raster referencing so now all rasters are X,Y
Created functions for linearising and delinearising points, for easier modifications
Added an effective midflame windspeed parameter that includes the effect of slope, according to the Rothermel model.
Changed some of the error messages and added some exceptions
Added a new exception to make the code fail if the closest the fire gets to the WUI area is more than 6 cells (the fire is deemed too far for a trigger boundary to form).